fix: dynamodb table creation
Although pulumi.DependsOn is used, the lambda permissions and
s3 bucket notificaton creation might not be synchronized perfectly
in terms of the timing of their application. This was causing the
notification setup to attempt validation before the permission is
fully active.

Add depends on for permission and also create dynamodb table for
aggregated values